Rename the import from `flickr_photos_api` to `flickr_api`
