+++ src/flickr_api/api/base.py
