+++ src/flickr_api/api/license_methods.py
