+++ src/flickr_api/exceptions.py
