+++ src/flickr_api/py.typed
