+++ src/flickr_api/types/__init__.py
