+++ src/flickr_api/types/licenses.py
