+++ src/flickr_api/types/machine_tags.py
