+++ src/flickr_api/types/sizes.py
