+++ tests/test_utils.py
@@ -1,10 +1,10 @@
 """
-Tests for ``flickr_photos_api.utils``.
+Tests for ``flickr_api.utils``.
 """
 
 import pytest
 
-from flickr_photos_api.utils import parse_safety_level
+from flickr_api.utils import parse_safety_level
 
 
 def test_unrecognised_safety_level_is_error() -> None:
