+++ tests/types/test_machine_tags.py
@@ -1,8 +1,8 @@
 """
-Tests for ``flickr_photos_api.types.machine_tags``.
+Tests for ``flickr_api.types.machine_tags``.
 """
 
-from flickr_photos_api.types import get_machine_tags
+from flickr_api.types import get_machine_tags
 
 
 def test_empty_tags_is_empty_machine_tags() -> None:
